refactor: reorganize config and utility modules
