+++ src/config/idealRangeConfig.js
