+++ src/constants/spectralColors.js
